refactor(extensions): ♻️ extract runtime tools into dedicated module
Move ResolvedTool, ToolProviderContext, and related tool resolution,
execution, and hook methods from extensions/mod.rs into a new
runtime_tools.rs module to improve separation of concerns and reduce
the size of the facade file.

refactor(workbench): ♻️ extract dashboard logic and thread surface state

- Move dashboard orchestration helpers and constants from
  dashboard-workbench.tsx into a dedicated dashboard-workbench-logic.ts
  module for better reuse and readability
- Extract runtime thread surface state, timeline mapping, and related
  types into a new runtime-thread-surface-state.ts module to slim down
  the main surface component
- Update imports and references across workbench-shell UI modules
- Update AGENTS.md to reflect the new module structure
