+++ AGENTS.md
@@ -1,7 +1,7 @@
 # Repository Guidelines
 
 ## Project Structure & Module Organization
-The desktop UI lives in `src/` with React + TypeScript. Use `src/app/` for bootstrap, routing, providers, and global styles; `src/modules/` for domain areas such as workbench, settings, marketplace, and the newer extensions center surfaces; `src/features/` for platform-facing features like terminal and system metadata; `src/shared/` for reusable UI, types, config, and helpers; and `src/services/` for bridge and streaming integrations. Static assets belong in `public/`. In `src/modules/workbench-shell/ui/`, keep large workbench surfaces as orchestrators and place extracted presentation helpers and small logic helpers beside them, such as `runtime-thread-surface-*`, `long-message-body.tsx`, `dashboard-sidebar.tsx`, `dashboard-overlays.tsx`, `dashboard-terminal-orchestrator.tsx`, `dashboard-workbench-logic.ts`, and `thread-rename-input.tsx`. Tauri/Rust code lives in `src-tauri/src/`: extension host/runtime code is organized under `src-tauri/src/extensions/` with `mod.rs` as the facade plus `config_io.rs`, `plugins.rs`, `mcp.rs`, `skills.rs`, and `marketplace.rs`; agent runtime orchestration stays under `src-tauri/src/core/` with focused companion modules such as `agent_session_execution.rs`, `agent_session_*`, `agent_run_compaction.rs`, and other `agent_run_*` modules. Migrations stay in `src-tauri/migrations/`, backend integration tests in `src-tauri/tests/`, and design references in `docs/`.
+The desktop UI lives in `src/` with React + TypeScript. Use `src/app/` for bootstrap, routing, providers, and global styles; `src/modules/` for domain areas such as workbench, settings, marketplace, and the newer extensions center surfaces; `src/features/` for platform-facing features like terminal and system metadata; `src/shared/` for reusable UI, types, config, and helpers; and `src/services/` for bridge and streaming integrations. Static assets belong in `public/`. In `src/modules/workbench-shell/ui/`, keep large workbench surfaces as orchestrators and place extracted presentation helpers and small logic helpers beside them, such as `runtime-thread-surface-*`, `runtime-thread-surface-state.ts`, `long-message-body.tsx`, `dashboard-sidebar.tsx`, `dashboard-overlays.tsx`, `dashboard-terminal-orchestrator.tsx`, `dashboard-workbench-logic.ts`, and `thread-rename-input.tsx`. Tauri/Rust code lives in `src-tauri/src/`: extension host/runtime code is organized under `src-tauri/src/extensions/` with `mod.rs` as the facade plus `config_io.rs`, `plugins.rs`, `mcp.rs`, `skills.rs`, `marketplace.rs`, and `runtime_tools.rs`; agent runtime orchestration stays under `src-tauri/src/core/` with focused companion modules such as `agent_session_execution.rs`, `agent_session_*`, `agent_run_compaction.rs`, and other `agent_run_*` modules. Migrations stay in `src-tauri/migrations/`, backend integration tests in `src-tauri/tests/`, and design references in `docs/`.
 
 ## Build, Test, and Development Commands
 - `npm run dev` — start the full Tauri desktop app.
